+++ CHANGELOG.rst
@@ -17,6 +17,9 @@ Version 0.1.1 (in development)
   elements that are not in the possibility space, instead of silently
   omitting them (fixes issue #3, reported by Erik Quaeghebeur).
 
+* LowProb.extend now has default upper=False argument, as it would
+  raise a ValueError instead.
+
 Version 0.1.0 (24 August 2010)
 ------------------------------